--- ProjectRRR/src/main/java/com/rudi/config/ConnectionManager.java
@@ -1,42 +1,0 @@
-package com.rudi.config;
-
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityManagerFactory;
-import jakarta.persistence.EntityTransaction;
-import jakarta.persistence.Persistence;
-
-import java.util.function.Consumer;
-
-public class ConnectionManager {
-    private EntityManagerFactory entityManagerFactory;
-    private EntityManager entityManager;
-
-    public ConnectionManager()
-    {
-        this.entityManagerFactory = Persistence.createEntityManagerFactory("persistenceUnitName");
-        this.entityManager = entityManagerFactory.createEntityManager();
-    }
-
-    public void inTransaction(Consumer<EntityManager> work) {
-        EntityTransaction transaction = entityManager.getTransaction();
-        try {
-            transaction.begin();
-            work.accept(entityManager);
-            transaction.commit();
-        }
-        catch (Exception e) {
-            if (transaction.isActive()) {
-                transaction.rollback();
-            }
-            throw e;
-        }
-        finally {
-            entityManager.close();
-        }
-    }
-
-    public EntityManager getEntityManager()
-    {
-        return this.entityManager;
-    }
-}
