+++ ProjectRRR/src/main/java/com/rudi/dao/AbstractDAO.java
@@ -1,26 +1,30 @@
 package com.rudi.dao;
 
-import com.rudi.config.ConnectionManager;
 import com.rudi.entities.AbstractEntity;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityManagerFactory;
+import jakarta.persistence.EntityTransaction;
+import jakarta.persistence.Persistence;
+
+import java.util.function.Consumer;
 
 public abstract class AbstractDAO {
-    protected ConnectionManager connectionManager;
 
     public AbstractDAO()
     {
-        this.connectionManager = new ConnectionManager();
+
     }
 
     public <T extends AbstractEntity> void save(T entity)
     {
-        connectionManager.inTransaction(session -> {
+        inTransaction(session -> {
             session.persist(entity);
         });
     }
 
     public <T extends AbstractEntity> void delete(T entity)
     {
-        connectionManager.inTransaction(session -> {
+        inTransaction(session -> {
             session.remove(session.merge(entity));
         });
     }
@@ -31,4 +35,23 @@ public abstract class AbstractDAO {
 //            session.find()
 //        });
 //    }
+    public void inTransaction(Consumer<EntityManager> work) {
+        EntityManagerFactory emf = Persistence.createEntityManagerFactory("persistenceUnitName");
+        EntityManager entityManager = emf.createEntityManager();
+        EntityTransaction transaction = entityManager.getTransaction();
+        try {
+            transaction.begin();
+            work.accept(entityManager);
+            transaction.commit();
+        }
+        catch (Exception e) {
+            if (transaction.isActive()) {
+                transaction.rollback();
+            }
+            throw e;
+        }
+        finally {
+            entityManager.close();
+        }
+    }
 }
